backend/requirements.txt
fastapi
uvicorn[standard]
python-socketio[asyncio]
cryptography
pqcrypto
